requirements.txt
beautifulsoup4==4.12.3
mysqlclient==2.2.4
pywikibot==9.1.2
PyYAML==6.0.1
requests==2.32.2
